fix: separate data store into mutable and immutable versions
