+++ .changeset/fresh-tigers-cheat.md
@@ -1,0 +1,5 @@
+---
+'astro': patch
+---
+
+Prevents content layer importing node builtins in runtime
