+++ packages/astro/src/content/content-layer.ts
@@ -12,12 +12,12 @@ import {
 	DATA_STORE_FILE,
 	MODULES_IMPORTS_FILE,
 } from './consts.js';
-import type { DataStore } from './data-store.js';
 import type { LoaderContext } from './loaders/types.js';
+import type { MutableDataStore } from './mutable-data-store.js';
 import { getEntryDataAndImages, globalContentConfigObserver, posixRelative } from './utils.js';
 
 export interface ContentLayerOptions {
-	store: DataStore;
+	store: MutableDataStore;
 	settings: AstroSettings;
 	logger: Logger;
 	watcher?: FSWatcher;
@@ -25,7 +25,7 @@ export interface ContentLayerOptions {
 
 export class ContentLayer {
 	#logger: Logger;
-	#store: DataStore;
+	#store: MutableDataStore;
 	#settings: AstroSettings;
 	#watcher?: FSWatcher;
 	#lastConfigDigest?: string;
