+++ packages/astro/src/content/loaders/types.ts
@@ -1,7 +1,7 @@
 import type { FSWatcher } from 'vite';
 import type { ZodSchema } from 'zod';
 import type { AstroIntegrationLogger, AstroSettings } from '../../@types/astro.js';
-import type { MetaStore, ScopedDataStore } from '../data-store.js';
+import type { MetaStore, ScopedDataStore } from '../mutable-data-store.js';
 
 export interface ParseDataOptions<TData extends Record<string, unknown>> {
 	/** The ID of the entry. Unique per collection */
